WIP: Refactoring using TRIBITS_INSTALL_HEADERS() (trilinos/Trilinos#1400)
This is what the kokkos/core/src/CMakeLists.txt file would look like if we
added TRIBITS_INSTALL_HEADERS() (see TriBITS backlog Google Doc).  That would
not be hard to implement on the TriBITS side and wouild avoid installing
header files when the user does not want to (i.e. when configuring with
-D<Project>_INSTALL_LIBRARIES_AND_HEADERS=OFF.)